requirements.txt
fastapi==0.115.12
google_api_python_client==2.166.0
google_auth_oauthlib==1.2.1
langchain==0.3.23
pandas==2.2.3
protobuf==4.25.3
pydantic==2.11.3
python-dotenv==1.1.0
Requests==2.32.3
uvicorn==0.34.0
